Fix netlify function
