+++ package.json
@@ -24,6 +24,7 @@
   "devDependencies": {
     "@tailwindcss/ui": "^0.6.2",
     "@types/faker": "^5.1.2",
+    "@types/prismjs": "^1.16.1",
     "@typescript-eslint/eslint-plugin": "^2.33.0",
     "@typescript-eslint/parser": "^2.33.0",
     "@vue/cli-plugin-babel": "~4.5.0",
@@ -62,8 +63,5 @@
   "repository": {
     "type": "git",
     "url": "git+https://github.com/rymate1234/tab-complete-input-vue.git"
-  },
-  "dependencies": {
-    "@types/prismjs": "^1.16.1"
   }
 }
