+++ tailwind.config.js
@@ -1,3 +1,3 @@
 module.exports = {
-  plugins: [require("@tailwindcss/ui")],
+  plugins: [require("@tailwindcss/ui")]
 };
